resolve conflicts, add gitignore file
