+++ .idea/uiDesigner.xml
@@ -1,4 +1,4 @@
-m<?xml version="1.0" encoding="UTF-8"?>
+<?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Palette2">
     <group name="Swing">
