+++ sprint1/src/Controller/CalculatorController.java
@@ -1,5 +1,8 @@
+package Controller;
+
 import Model.CalculatorModel;
 import View.CalculatorView;
+
 import java.util.Scanner;
 
 import java.util.Scanner;
