+++ sprint1/src/Controller/UserInputController.java
@@ -1,3 +1,5 @@
+package Controller;
+
 import java.util.Scanner;
 
 public class UserInputController {
